+++ src/pages/MyOrderPage.tsx
@@ -1,27 +1,17 @@
 import { useMemo, useState } from 'react';
-import { useLocation, useNavigate } from 'react-router-dom';
-import Button from '../components/common/Button';
-import ProductCard from '../components/common/ProductCard';
+import { useLocation } from 'react-router-dom';
+import GlobalModal from '../components/common/GlobalModal';
+import MyOrderInfo from '../components/MyOrderInfo';
+import MyOrderList from '../components/MyOrderList';
+import MyReviewModal from '../components/MyReviewModal';
 import useIntersectHandler from '../hooks/useIntersectHandler';
-import useViewport from '../hooks/useViewport';
 import useGetOrderQuery from '../query/useGetOrderQuery';
-import theme from '../shared/style/theme';
-import type { IList, TMyAPIResOrder } from '../shared/types/types';
+import type { IList, TResOrder } from '../shared/types/types';
 import * as t from '../style/myOrderPage.style';
-// import GlobalModal from './common/GlobalModal';
-// import MyReviewModal from './MyReviewModal';
-
-const getList = (data: TMyAPIResOrder): IList[] => {
-  if ('orderList' in data) return data.orderList;
-  else return data.cancelList;
-};
 
 export default function MyOrderPage() {
   const pathPattern = useLocation();
   const [, path] = pathPattern.pathname.split('/');
-
-  const navigate = useNavigate();
-  const viewport = useViewport();
 
   const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
 
@@ -34,6 +24,10 @@ export default function MyOrderPage() {
     hasNextPage,
   } = useGetOrderQuery(path);
 
+  const getList = (list: TResOrder): IList[] => {
+    if ('orderList' in list) return list.orderList;
+    else return list.cancelList;
+  };
   /** Data Filtering */
   const orderData = useMemo(
     () => data?.pages.map(page => getList(page?.data)).flat() || null,
@@ -46,72 +40,31 @@ export default function MyOrderPage() {
     fetchNextPage();
   });
 
-  const reviweBtnClick = () => setIsModalOpen(!isModalOpen);
   const reviewModal = isModalOpen && (
-    // <GlobalModal onClose={() => setIsModalOpen(false)}>
-    //   <MyReviewModal />
-    // </GlobalModal>
-    <></>
+    <GlobalModal onClose={() => setIsModalOpen(false)}>
+      <MyReviewModal onClose={() => setIsModalOpen(false)} />
+    </GlobalModal>
   );
-  const cancelBtnClick = (item: IList) => navigate(`cancel-call/${item.o_No}`);
 
   if (isLoading) return <p>Loading...</p>;
-  if (orderData.length === 0)
-    return <t.DataNull>주문 내역이 없습니다.</t.DataNull>;
-
   return (
-    <t.Section>
+    <t.Base>
       {reviewModal}
       <t.Title>주문조회</t.Title>
       <t.Article>
+        {orderData.length === 0 && (
+          <t.DataNull>주문 내역이 없습니다.</t.DataNull>
+        )}
         {orderData.map((list, i) => (
           <t.OrderList key={i}>
-            <t.TopInfo>
-              <t.OrderNumber>
-                <t.Label> 주문번호 </t.Label>
-                <t.ProductLink to={`${list.o_No}`} className="link">
-                  <span>
-                    {viewport > 990 ? `${list.o_No}` : '주문상세보기'}
-                  </span>
-                  <t.ArrowIcon />
-                </t.ProductLink>
-              </t.OrderNumber>
-              <t.Date>
-                <t.Label> 주문일자 </t.Label>
-                <span>{list.o_Date}</span>
-              </t.Date>
-            </t.TopInfo>
-            <t.Main>
-              {list.products.map((item: any, idx: number) => (
-                <t.Container key={idx}>
-                  <t.ProductWrapper>
-                    <t.ProductLink to={`${list.o_No}`} className="link">
-                      <ProductCard
-                        p_Thumbnail={item.p_Thumbnail}
-                        a_Brand={item.a_Brand}
-                        p_Name={item.p_Name}
-                        p_Option={item.p_Option}
-                      />
-                      <t.ProductStatus>{item.o_Status}</t.ProductStatus>
-                    </t.ProductLink>
-                  </t.ProductWrapper>
-                  <t.ActionButtonContainer>
-                    {item.o_Status === '주문완료' && (
-                      <Button
-                        {...cancelStyle}
-                        onClick={() => cancelBtnClick(list)}
-                      ></Button>
-                    )}
-                    {item.p_Status && item.r_Status && (
-                      <Button
-                        {...reviweStyle}
-                        onClick={() => reviweBtnClick()}
-                      ></Button>
-                    )}
-                  </t.ActionButtonContainer>
-                </t.Container>
-              ))}
-            </t.Main>
+            <MyOrderInfo list={list} />
+            <MyOrderList
+              products={list.products}
+              orderNo={list.o_No}
+              isModalOpen={isModalOpen}
+              setIsModalOpen={setIsModalOpen}
+              type="order"
+            />
           </t.OrderList>
         ))}
       </t.Article>
@@ -120,27 +73,6 @@ export default function MyOrderPage() {
       ) : (
         <div ref={target}></div>
       )}
-    </t.Section>
+    </t.Base>
   );
 }
-
-const cancelStyle = {
-  bgColor: 'transparent',
-  radius: '30px',
-  border: `1px solid ${theme.rgba04}`,
-  hBgColor: 'transparent',
-  hBorder: `1px solid ${theme.rgba08}`,
-  hColor: `${theme.fc14}`,
-  color: `${theme.fc14}`,
-  fontSize: `${theme.fc12}`,
-  fontWeight: '500',
-  padding: '10px 16px',
-  text: '취소',
-};
-const reviweStyle = {
-  radius: '30px',
-  fontSize: `${theme.fs12}`,
-  fontWeight: '500',
-  padding: '10px 16px',
-  text: '구매평 작성',
-};
